TuuzPC

+++ app/v1/iflytek/model/IflytekModel.py
@@ -24,24 +24,24 @@ class IflytekModel(BaseModel):
         super().__init__(conn)
 
     def api_find_byId(self, id):
-        return Database.Db().table(self.Table).where("id", id).find()
+        return Database.Db(self.db).table(self.Table).where("id", id).find()
 
     def api_insert(self, uid, name, team_id, host, api_secret, app_id, api_key, vcn):
-        return Database.Db().table(self.Table).insert(
+        return Database.Db(self.db).table(self.Table).insert(
             {"uid": uid, "name": name, "team_id": team_id, "host": host, "api_secret": api_secret, "app_id": app_id, "api_key": api_key, "vcn": vcn})
 
     def api_update_byUidAndId(self, uid, id, name, team_id, host, api_secret, app_id, api_key, vcn):
-        return Database.Db().table(self.Table).where("uid", uid).where("id", id).update(
+        return Database.Db(self.db).table(self.Table).where("uid", uid).where("id", id).update(
             {"name": name, "team_id": team_id, "host": host, "api_secret": api_secret, "app_id": app_id, "api_key": api_key, "vcn": vcn})
 
     def api_delete(self, id):
-        return Database.Db().table(self.Table).where("id", id).delete()
+        return Database.Db(self.db).table(self.Table).where("id", id).delete()
 
     def api_delete_byUidAndId(self, uid, id):
-        return Database.Db().table(self.Table).where("uid", uid).where("id", id).delete()
+        return Database.Db(self.db).table(self.Table).where("uid", uid).where("id", id).delete()
 
     def api_find_byUidAndId(self, uid, id):
-        return Database.Db().table(self.Table).where("uid", uid).where("id", id).find()
+        return Database.Db(self.db).table(self.Table).where("uid", uid).where("id", id).find()
 
     def api_select_byUid(self, uid):
-        return Database.Db().table(self.Table).where("uid", uid).select()
+        return Database.Db(self.db).table(self.Table).where("uid", uid).select()
